all: Add EU_863_870 band definition
